gh #42: working in JAXB to render sonarcube info
This isn't wired into the runtime of the project yet but the problem we have is Cobertura doesn't reference the original sources directory for multiple source sets in it's XML output.  That being said, in the ProjectData model object we get the filename in the form "my/package/MyClass.java" from the ClassData.getSourceFileName() which we can use as the basis for testing all sourcesPaths to see if in which one a given file exists.  With that then we can generate a proper Sonarqube XML representation by just rendering out the JAXB objects to the sonarqube xml output file.